server/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
httpx
python-dotenv
pytest
pytest-asyncio
vnstock
feedparser
sqlalchemy
perplexityai
newspaper3k
lxml[html_clean]
apscheduler